lighttpd: Patch lighttpd to use custom pcre lib
Implemented a patch to the lighttpd CMake build system so that
it can support a custom libprce config script. Prior the libpcre
config script would only be sourced from the sysroot that lighttpd
is being built on. This patch allows lighttpd to take an
externally provided pcre config and library that is not installed
in the users sysroot. Additionally updated the external project
to checkout lighttpd on a specific tag (so that the patch can
be reliably applied)