multi_main.camkes: A multi-vm webserver variant
Added a variant camkes application of the webserver. This
variant establishes multiple VM's, each running a webserver. In
addition there exists a VM with ethernet passthrough that in turn
faciliates and multiplexes external network access for the other
VM's. The VM with ethernet access provides a virtual network
interface to the other VM's so that they can access the external
network.