lighttpd: Patch lighttpd to compile lemon on host
Implemented a patch to the lighttpd CMake build system so that
it compiles the lemon binary with the hosts toolchain. This is
needed to ensure the host machine can execute the lemon
binary when building lighttpd. This would otherwise cause issues
when cross-compiling lighttpd, producing a lemon binary for a
different architecture.